+++ wifly_thread.h
@@ -7,7 +7,7 @@
 int wifly_connect(char *port);
 
 /* update the booleans that keep track of the current state */
-void update_state(int &new_state);
+void update_state(uint8_t &new_state);
 
 /* main thread function to be called by pthread in mod.cpp */
 void *wifly_thread(void *param);
